[REF] (*)dropshipping: separate from Sale module
Allow dropshipping in inventory and mrp apps
without installing the Sale app.

task: 3229537
