+++ addons/purchase_stock/models/res_config_settings.py
@@ -10,11 +10,3 @@ class ResConfigSettings(models.TransientModel):
     module_stock_dropshipping = fields.Boolean("Dropshipping")
     days_to_purchase = fields.Float(
         related='company_id.days_to_purchase', readonly=False)
-    is_installed_sale = fields.Boolean(string="Is the Sale Module Installed")
-
-    def get_values(self):
-        res = super(ResConfigSettings, self).get_values()
-        res.update(
-            is_installed_sale=self.env['ir.module.module'].search([('name', '=', 'sale'), ('state', '=', 'installed')]).id,
-        )
-        return res
