+++ addons/sale_mrp_subcontracting_dropshipping/models/__init__.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import purchase
-from . import res_company
-from . import stock
+from . import stock_picking
