+++ addons/sale_mrp_subcontracting_dropshipping/tests/__init__.py
@@ -1,4 +1,5 @@
 #  -*- coding: utf-8 -*-
 #  Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import test_anglo_saxon_valuation
+from . import test_purchase_subcontracting
+from . import test_sale_dropshipping
