+++ addons/sale_stock_dropshipping/__init__.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import purchase
-from . import res_company
-from . import stock
+from . import models
