+++ addons/sale_stock_dropshipping/data/stock_data.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<odoo>
+    <data noupdate="1">
+        <record id="route_drop_shipping" model='stock.route'>
+            <field name="sale_selectable" eval="True"/>
+        </record>
+    </data>
+</odoo>
