+++ addons/sale_stock_dropshipping/models/__init__.py
@@ -2,5 +2,5 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import purchase
-from . import res_company
+from . import sale
 from . import stock
