+++ addons/sale_stock_dropshipping/tests/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from . import test_crossdock
+from . import test_dropship
+from . import test_procurement_exception
+from . import test_stockvaluation
