+++ addons/sale_stock_dropshipping/tests/test_dropship.py
