+++ addons/sale_stock_dropshipping/views/sale_order_views.xml
