+++ addons/stock_dropshipping/data/stock_data.xml
@@ -8,13 +8,10 @@
             <field name="name">Dropship</field>
             <field name="sequence">20</field>
             <field name="company_id"></field>
-            <field name="sale_selectable" eval="True"/>
             <field name="product_selectable" eval="True"/>
             <field name="product_categ_selectable" eval="True"/>
         </record>
 
-        <function model="res.company" name="create_missing_dropship_sequence"/>
-        <function model="res.company" name="create_missing_dropship_picking_type"/>
         <function model="res.company" name="create_missing_dropship_rule"/>
     </data>
 </odoo>
